requirements.txt
PySimpleGUI==4.60.5
Pillow==10.1.0
PyPDF2==3.0.1
reportlab==4.0.8
pdf2image==1.16.3
